requirements.txt
torch
torchvision
numpy
PyYAML
easydict
tensorboardX
opencv-python
Pillow
scikit-learn
pytorch-metric-learning
tabulate
setproctitle
